Setup react-navigation
