+++ playground/src/stories/ErrorScreen.js
@@ -4,7 +4,7 @@ import { ErrorScreen as ES, Text, Button, injectTheme } from 'rn-hero-design';
 const ErrorScreen = ({ theme }) => (
   <ES
     onPressClose={() => {}}
-    imageSource={require('../assets/404_icon.png')}
+    imageSource={require('../images/404_icon.png')}
     title={
       <Text size="h1" style={{ textAlign: 'center' }}>
         Oops! Something has been wrong
@@ -14,8 +14,7 @@ const ErrorScreen = ({ theme }) => (
       <Text
         size="h4"
         color={theme.variables.MUTED_TEXT_COLOR}
-        style={{ textAlign: 'center' }}
-      >
+        style={{ textAlign: 'center' }}>
         If you keep encounting this issue, please contact with our{' '}
         <Text weight="bold" color={theme.variables.MUTED_TEXT_COLOR}>
           Support team
